test: move valkey incremental reply scanning into its own file
